令牌桶开发

+++ ddd-akka-ratelimiter/src/main/java/com/zero/ddd/akka/ratelimiter/state/vo/RateLimiterConfig.java
@@ -18,6 +18,6 @@ import lombok.NoArgsConstructor;
 public class RateLimiterConfig {
 	
 	private String rateLimiterName;
-	private int permitsPerSecond;
+	private double permitsPerSecond;
 
 }
